mvp/pyproject.toml
[project]
name = "mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ryanair-py @ file:/home/masaccio/lesgoski/mvp/repos/ryanair-py",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.53.1",
    "tls-client>=1.0.1",
]
